Refactor the TCK serialization of results
