+++ community/cypher/compatibility-suite/src/main/java/cypher/feature/parser/matchers/ResultWrapper.java
@@ -88,107 +88,9 @@ public class ResultWrapper implements Result
         Map<String,Object> next = inner.next();
         builder.append( "[" ).append( rowCounter++ ).append( "]   " );
         builder.append( "actualRow:" );
-        builder.append( serialize( next ) );
+        builder.append( TckSerializer.serialize( next ) );
         builder.append( "\n" );
         return next;
-    }
-
-    private String serialize( Object obj )
-    {
-        StringBuilder sb = new StringBuilder();
-        if ( obj == null )
-        {
-            sb.append( "null" );
-        }
-        else if ( obj instanceof Node )
-        {
-            Node n = (Node) obj;
-            sb.append( "(" );
-            n.getLabels().forEach( ( l ) -> sb.append( ":" ).append( l.name() ) );
-            sb.append( " {" );
-            String[] comma = new String[]{ "" };
-            n.getAllProperties().forEach( ( k, v ) -> {
-                sb.append( comma[0] ).append( k ).append( ": " ).append( serialize( v ) );
-                comma[0] = ", ";
-            } );
-            sb.append( "})" );
-        }
-        else if ( obj instanceof Relationship )
-        {
-            Relationship r = (Relationship) obj;
-            sb.append( "[:" );
-            sb.append( r.getType().name() );
-            sb.append( " {" );
-            String[] comma = new String[]{ "" };
-            r.getAllProperties().forEach( ( k, v ) -> {
-                sb.append( comma[0] ).append( k ).append( ": " ).append( serialize( v ) );
-                comma[0] = ", ";
-            } );
-            sb.append( "}]" );
-        }
-        else if ( obj instanceof Path )
-        {
-            Path p = (Path) obj;
-            sb.append( "<" );
-            sb.append( Paths.pathToString( p, new Paths.PathDescriptor<Path>()
-            {
-                @Override
-                public String nodeRepresentation( Path path, Node node )
-                {
-                    return serialize( node );
-                }
-
-                @Override
-                public String relationshipRepresentation( Path path,
-                        Node from, Relationship relationship )
-                {
-                    String prefix = "-", suffix = "-";
-                    if ( from.equals( relationship.getEndNode() ) )
-                    {
-                        prefix = "<-";
-                    }
-                    else
-                    {
-                        suffix = "->";
-                    }
-                    return prefix + serialize( relationship ) + suffix;
-                }
-            } ) );
-            sb.append( ">" );
-        }
-        else if ( obj instanceof String )
-        {
-            sb.append( "'" ).append( obj.toString() ).append( "'" );
-        }
-        else if ( obj instanceof List )
-        {
-            List<?> list = (List) obj;
-            List<String> output = new ArrayList<>( list.size() );
-            list.forEach( item -> output.add( serialize( item ) ) );
-            sb.append( output );
-        }
-        else if ( obj.getClass().isArray() )
-        {
-            List<Object> list = new ArrayList<>();
-            for ( int i = 0; i < Array.getLength( obj ); ++i )
-            {
-                list.add( Array.get( obj, i ) );
-            }
-            sb.append( serialize( list ) );
-        }
-        else if ( obj instanceof Map )
-        {
-            Map<?,?> map = (Map) obj;
-            Map<String,String> output = new HashMap<>( map.size() );
-            map.forEach( ( k, v ) -> output.put( k.toString(), serialize( v ) ) );
-            sb.append( output );
-        }
-        else
-        {
-            sb.append( obj.toString() );
-        }
-
-        return sb.toString();
     }
 
     @Override
